chore: refactor Equipment subclasses
