+++ src/game/equipments/armour/Chainmail.java
@@ -1,0 +1,10 @@
+package game.equipments.armour;
+
+import game.equipments.Equipment;
+import game.equipments.EquipmentType;
+
+public class Chainmail extends Equipment {
+    public Chainmail() {
+        super("Chainmail", EquipmentType.Armour, 70, 0, 1, 0, -1);
+    }
+}
