+++ src/game/equipments/armour/Fleece.java
@@ -1,0 +1,10 @@
+package game.equipments.armour;
+
+import game.equipments.Equipment;
+import game.equipments.EquipmentType;
+
+public class Fleece extends Equipment {
+    public Fleece() {
+        super("Fleece", EquipmentType.Armour, 150, 0, 2, 1, -1);
+    }
+}
