+++ src/game/equipments/armour/Regalia.java
@@ -1,0 +1,10 @@
+package game.equipments.armour;
+
+import game.equipments.Equipment;
+import game.equipments.EquipmentType;
+
+public class Regalia extends Equipment {
+    public Regalia() {
+        super("Regalia", EquipmentType.Armour, 105, 0, 1, 0, 0);
+    }
+}
