+++ src/game/equipments/artefact/Amulet.java
@@ -1,0 +1,10 @@
+package game.equipments.artefact;
+
+import game.equipments.Equipment;
+import game.equipments.EquipmentType;
+
+public class Amulet extends Equipment {
+    public Amulet() {
+        super("Amulet", EquipmentType.Artefact, 200, 1, -1, 1, 1);
+    }
+}
