+++ src/game/equipments/artefact/Crystal.java
@@ -1,0 +1,10 @@
+package game.equipments.artefact;
+
+import game.equipments.Equipment;
+import game.equipments.EquipmentType;
+
+public class Crystal extends Equipment {
+    public Crystal() {
+        super("Crystal", EquipmentType.Artefact, 210, 2, 1, -1, -1);
+    }
+}
