+++ src/game/equipments/artefact/Excalibur.java
@@ -1,0 +1,10 @@
+package game.equipments.artefact;
+
+import game.equipments.Equipment;
+import game.equipments.EquipmentType;
+
+public class Excalibur extends Equipment {
+    public Excalibur() {
+        super("Excalibur", EquipmentType.Artefact, 150, 2, 0, 0, 0);
+    }
+}
